finish second and third scenario
The second lets you click a new link on a movie details page
"Find Movies With Same Director", and shows all movies that share
the same director as the displayed movie.

The third handles the sad path, when the current movie
has no director info but we try to do "Find with same director" anyway.